requirements.txt
playwright>=1.49.0
python-dotenv>=1.0.1
pandas
openpyxl
fastapi
uvicorn[standard]
slack-sdk
aiohttp
sqlmodel
sqlalchemy
alembic
bcrypt
itsdangerous
python-multipart
ruff
